+++ client-core/src/main/java/com/influxdb/internal/UserAgentInterceptor.java
@@ -1,0 +1,54 @@
+/*
+ * The MIT License
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+package com.influxdb.internal;
+
+import java.io.IOException;
+import javax.annotation.Nonnull;
+
+import okhttp3.Interceptor;
+import okhttp3.Request;
+import okhttp3.Response;
+
+/**
+ * @author Jakub Bednar (02/03/2020 10:10)
+ */
+public class UserAgentInterceptor implements Interceptor {
+
+    private final String userAgent;
+
+    public UserAgentInterceptor() {
+        Package mainPackage = UserAgentInterceptor.class.getPackage();
+        String version = mainPackage.getImplementationVersion();
+
+        userAgent = "influxdb-client-java/" + (version != null ? version : "unknown");
+    }
+
+    @Nonnull
+    @Override
+    public Response intercept(final Chain chain) throws IOException {
+
+        Request request = chain.request();
+        Request.Builder builder = request.newBuilder().header("User-Agent", userAgent);
+
+        return chain.proceed(builder.build());
+    }
+}